backend/requirements.txt
# API
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6
httpx>=0.26.0

# AI (Gemini) — use `google.genai`, not deprecated `google-generativeai`
google-genai>=1.0.0

# Regulatory RAG (local Chroma + embeddings + PDF extract)
chromadb>=0.4.22
sentence-transformers>=2.3.0
PyMuPDF>=1.23.0

# PDF export
reportlab>=4.0.0
